Update the MessageDecoder to take a CamusWrapper to set decoded value
The pattern of setting a deserialized value in a given reference instead
of returning a newly allocated object is typical in hadoop. It's what is
done with the key in EtlRecordReader and now this establishes the same
for the record. It basically allows us to avoid allocating a new hashmap
for every decoded value.