Refactor EtlKey class and how it's handled in KafkaReader
This CL cleans up the EtlKey class a bunch, including some legacy stuff
from linkedin we don't care about. I removed the big "set everything"
method and added setters for the individual fields; it isn't necessary
to set everything for each message, since the topic, leaderId, partition
and presumably beginningOffset won't change for each reader. I'm
actually not sure about that last one, because it was being updated to
the previous offset with each request before. Now we just leave at the
initial beginning offset value, which is set from the EtlRequest that
initializes the reader.