Fix request offset resetting; refactor and clean up EtlRequest
This fixes the logic where an offset request is outside the bounds of
the available earliest and latest offsets for a partition. I had messed
it up in refactoring (though it was already suspect in it's behavior).
In the case where the offset is greater than the latest available, it
just sets it to the latest, which will probably mean that no data is
fetched. In the original code, it would set it to the *earliest*
available in that case, which is surely wrong. Maybe it's best to just
throw an exception.

Also made EtlREquest implement Comparable using topic and partition
number which will result in more consistent sorting (those are also the
fields used for equals() and hashcode()). Made the simple consumer use
the context to get configuration values for timeouts etc, rather than
hard coding them.

Also, when the getLatestOffset(long) is called directly (and it's
corollary for earliest), the cached value isn't changed directly, since
that should presumably be set by the no-arg method.