Big refactoring commit, lots of changes including EtlMultiOutputFormat
This CL started as an exploration into EtlMultiOutputFormat to determine
why camus was exhibiting some inconsistent behavior with regard to
starting a new job based on the offsets of the last.

EtlMultiOutputFormat contains lots of changes. The count tracking it was
doing was totally broken ('[]', and empty json string was being written
in the hdfs records). I ripped most of it out and added something
simple count tracking.

I added a KafkaBrokerClient to encapsulate many of the operations done
with SimpleConsumers, which the broker client takes as a collaborator.
This allowed cleanup in KafkaReader and EtlInputFormat.

Other changes in KafkaReader.java
  - use guava's AbstractIterator to help simplify interface
  - removed context (was only used to create simple consumer)
  - cleaned up stuff about skipping messages
  - renamed variables to make more clear (initial/final instead of
    begin/end

I also removed the context from the EtlRequest class, which is really
just a model class. It shouldn't create consumers and network
connections internally. The methods to get offsets are now in the client
broker.

Removed some other now (and previously) unused classes.

Probably more cleanup...