Add JobRecordKeeper interface for abstracting job meta-data tracking
This moves the code for writing offsets, requests (which are never read
back in the code), and counts (which we temporarily don't do, and didn't
work before anyway) into a new record keeping interface. Long term, we
want all mapreduce record keeping to be stored in hbase for better
realtime access and querying.

I updated the PartitionCounts class to maintain a separately allocated
EtlKey object which copies the changing values of the key when
incrementing counts. This prevents any strange things that might happen
with the passed in key used for writing the last entry from changing the
final offsets.

Fixed a bug where the same instance of EtlKey used to read the previous
offsets was inserted in the map (potentially multiple times). Now a new
instance is created whenever inserting a value into the map.

I also removed some more dead record counting code, and simplified the
concurrent map put-if-absent pattern. The contract for the method from
the Map interface is confusing for the concurrent version.

Change-Id: I584f5514da98500696a3b3993f1faef852b372aa